dom/tests/mochitest/dom-level0/file_location.html

<html>
    <script>
        try {
            parent.SimpleTest.ok(!("existingprop" in location), "got a new location object in the iframe");
        } catch (e) {
        }

        location.iframeprop = 42;
    </script>
</html>
